fix: caches work on unknown data types
the previous generic typing caused a situation where a single cache instance
could not be reused on different cachified calls operating on varying data types

This was never a design goal and happened as a over-optimization when implementing against
tests instead of a rl-app 😅

fix: https://github.com/Xiphe/cachified/issues/5

BREAKING CHANGE:
The `Cache` type is not generic anymore and always uses `unknown` for values
You might need to adjust your cache implementations and caches to now work on
unknown data types. Typescript should inform you where.
Nothing to be done when you only used the build-in cache adapters.